Fix clients page having no data
